pyproject.toml
[tool.poetry]
name = "django-xml"
version = "0.1.0"
description = "A Django app to process sitemaps and extract content."
authors = ["Emmanuel Oreoluwa <awuloero13@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
celery = "^5.4.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.3"
lxml = "^5.3.1"
defusedxml = "^0.7.1"
django-cleanup = "^9.0.0"
djangorestframework = "^3.15.2"
python-dateutil = "^2.9.0.post0"
playwright = "^1.50.0"
httpx = "^0.28.1"
python-dotenv = "^1.0.1"
redis = "^5.2.1"
django-jsonfield = "^1.4.1"
channels = "^4.2.0"
channels-redis = "^4.2.1"
daphne = "^4.1.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.10.0"
pytest-celery = "^1.2.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
